Don't bother writing out a page if it doesn't need it. This may be a stupid way of checking that, but hey.

+++ lib/salt/page.rb
@@ -33,11 +33,16 @@ module Salt
     end
 
     def write(site, path, context = {}, extension = nil)
-      contents = self.render(site, @contents, {this: self}.merge(context))
-      output_path = self.output_path(site, path)
 
-      FileUtils.mkdir_p(output_path) unless Dir.exists?(output_path)
+      output_path = self.output_path(site, path)
       full_path = File.join(output_path, self.output_file(extension))
+
+      if @path && File.exists?(full_path)
+        return unless File.mtime(@path) > File.mtime(full_path)
+      end
+
+      contents = self.render(site, @contents, {this: self}.merge(context))
+      FileUtils.mkdir_p(output_path) unless Dir.exists?(output_path)
 
       File.open(full_path, 'w') do |file|
         file.write(contents)
